consolidated to one EmptyStatus component, removed margin to remove double scrolling
