+++ src/ui/views/NftEvm/ListTab.tsx
@@ -17,7 +17,7 @@ import { useHistory } from 'react-router-dom';
 import { useWallet } from '@/ui/utils/WalletContext';
 import placeholder from 'ui/FRWAssets/image/placeholder.png';
 
-import EmptyStatus from './EmptyStatus';
+import EmptyStatus from '../EmptyStatus';
 
 interface ListTabProps {
   data: any;
